% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{adcc_copula_gradient}
\alias{adcc_copula_gradient}
\title{ADCC Copula Gradient (Numerical)}
\usage{
adcc_copula_gradient(
  params,
  z_matrix,
  weights,
  Qbar,
  Nbar = NULL,
  copula_dist = "mvn"
)
}
\arguments{
\item{params}{Parameter vector: (alpha, gamma, beta) or with shape for MVT}

\item{z_matrix}{Matrix of copula residuals}

\item{weights}{Observation weights}

\item{Qbar}{Unconditional covariance matrix}

\item{Nbar}{Unconditional covariance of negative shocks}

\item{copula_dist}{Copula distribution}
}
\value{
Gradient vector
}
\description{
Computes the gradient of ADCC copula NLL using numerical differentiation.
}
\keyword{internal}
