% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{add_boundary_event}
\alias{add_boundary_event}
\title{Add Boundary Event}
\usage{
add_boundary_event(
  diagnostics,
  iteration,
  state,
  parameter_name,
  value,
  boundary_type,
  action_taken
)
}
\description{
Add Boundary Event
}
\keyword{internal}
