% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{add_diagnostic_warning}
\alias{add_diagnostic_warning}
\title{Add Warning}
\usage{
add_diagnostic_warning(
  diagnostics,
  iteration,
  warning_type,
  message,
  details = NULL
)
}
\description{
Add Warning
}
\keyword{internal}
