% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{add_em_iteration_diagnostic}
\alias{add_em_iteration_diagnostic}
\title{Add EM Iteration Diagnostics}
\usage{
add_em_iteration_diagnostic(
  diagnostics,
  iteration,
  log_lik_before,
  log_lik_after,
  ll_change,
  duration_sec,
  parameters,
  convergence_flag = FALSE
)
}
\description{
Add EM Iteration Diagnostics
}
\keyword{internal}
