% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{add_parameter_evolution}
\alias{add_parameter_evolution}
\title{Add Parameter Evolution Data}
\usage{
add_parameter_evolution(diagnostics, iteration, state, parameters)
}
\description{
Add Parameter Evolution Data
}
\keyword{internal}
