% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{add_sigma_evolution}
\alias{add_sigma_evolution}
\title{Add Sigma Evolution Data}
\usage{
add_sigma_evolution(diagnostics, iteration, state, series, sigma_summary)
}
\description{
Add Sigma Evolution Data
}
\keyword{internal}
