% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_diagnostics.R
\name{as.data.frame.tsbs_diagnostics}
\alias{as.data.frame.tsbs_diagnostics}
\title{Convert Diagnostics to Data Frame}
\usage{
\method{as.data.frame}{tsbs_diagnostics}(x, ..., what = c("stats", "blocks", "all"))
}
\arguments{
\item{what}{Character specifying what to convert: "blocks", "stats", or "all".}

\item{diagnostics}{A \code{tsbs_diagnostics} object.}
}
\value{
A data frame.
}
\description{
Convert Diagnostics to Data Frame
}
