% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockBootstrap.R
\name{blockBootstrap}
\alias{blockBootstrap}
\title{Block Bootstrap for Time Series (Rcpp Implementation)}
\usage{
blockBootstrap(
  x,
  n_boot = NULL,
  block_length = NULL,
  bs_type = "moving",
  block_type = "overlapping",
  taper_type = "cosine",
  tukey_alpha = 0.5,
  num_blocks = NULL,
  num_boots = 20L,
  p = 0.1,
  stationary_max_percentile = 0.99,
  stationary_max_fraction_of_n = 0.1
)
}
\arguments{
\item{x}{A numeric matrix with rows representing time points and columns
representing variables.}

\item{n_boot}{Integer or \code{NULL}. Desired number of time points in the
bootstrap sample. If \code{NULL}, the original series length is used.}

\item{block_length}{Integer or \code{NULL}. Length of each block. If \code{NULL}, a
heuristic is used based on the average lag-1 autocorrelation across
variables.}

\item{bs_type}{Bootstrap type. Character string: One of \code{"moving"} or
\code{"stationary"}. Default \code{"moving"}.}

\item{block_type}{Character. Block type. Either \code{"non-overlapping"},
\code{"overlapping"} or \code{"tapered"}. Default \code{"overlapping"}.}

\item{taper_type}{Tapering window function. Character. One of \code{"cosine"},
\code{"bartlett"}, or \code{"tukey"}. Default \code{"cosine"}.}

\item{tukey_alpha}{\code{alpha} when \code{block_type = "tapered"} and  \code{taper_type = "tukey"}.
Default 0.5.}

\item{num_blocks}{Integer or \code{NULL}. Number of blocks to use. If \code{NULL}, this
is inferred from \code{n_boot}.}

\item{num_boots}{Number of bootstrap replicates to generate.}

\item{p}{numeric \eqn{p \in (0, 1)}. Probability parameter for the geometric
block length (used in Stationary Bootstrap). Default is 0.1.}

\item{stationary_max_percentile}{Stationary max percentile. Default is 0.99.}

\item{stationary_max_fraction_of_n}{Stationary max fraction of n. Default is
0.10.}
}
\value{
A list of matrices, each one a bootstrap replicate with dimensions
approximately \code{n_length} × \code{ncol(x)}.
}
\description{
Generates bootstrap replicates of a multivariate time series using either the
Moving Block Bootstrap (MBB) or Stationary Bootstrap (SB) methods.
Not intended to be called directly by user, only to be called internally by
tsbs().
}
\details{
The Moving Block Bootstrap (MBB) resamples fixed-length overlapping
blocks from the original time series. The Stationary Bootstrap (SB) samples
blocks with geometrically distributed lengths and can optionally enforce
non-overlap constraints.

If \code{n_boot} is set, the last block will be trimmed when necessary.
If \code{n_boot} is not set, and \code{num_blocks} is set, the length of each
bootstrap series will be determined by the number of blocks and the
lengths of the individual blocks for that particular series.
If neither \code{n_boot} nor \code{num_blocks} is set, \code{n_boot} will default to the
number of rows in \code{x} and the last block will be trimmed when necessary.
}
\examples{
set.seed(123)
x <- matrix(rnorm(100), ncol = 1)
boots <- blockBootstrap(x, bs_type = "stationary", num_boots = 5L)
dim(boots[[1]])

}
\references{
Politis, D., & Romano, J. (1994). The Stationary Bootstrap. Journal of the
American Statistical Association, 89, 1303-1313.
\url{http://dx.doi.org/10.1080/01621459.1994.10476870}
}
