% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_se.R
\name{cgarch_nll_for_hessian}
\alias{cgarch_nll_for_hessian}
\title{CGARCH Negative Log-Likelihood for Hessian Computation}
\usage{
cgarch_nll_for_hessian(
  params,
  z_matrix,
  weights,
  Qbar,
  copula_dist = "mvn",
  use_reparam = FALSE
)
}
\description{
Wrapper around copula NLL for use with numerical_hessian.
}
\keyword{internal}
