% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_se.R
\name{cgarch_sandwich_se}
\alias{cgarch_sandwich_se}
\title{CGARCH Sandwich (Robust) Standard Errors}
\usage{
cgarch_sandwich_se(
  params,
  z_matrix,
  weights,
  Qbar,
  copula_dist = "mvn",
  use_reparam = FALSE
)
}
\arguments{
\item{params}{Parameter vector}

\item{z_matrix}{Copula residuals}

\item{weights}{Observation weights}

\item{Qbar}{Unconditional covariance}

\item{copula_dist}{"mvn" or "mvt"}

\item{use_reparam}{Use reparameterization?}
}
\value{
List with SE results
}
\description{
Computes sandwich estimator for CGARCH parameters.
}
\keyword{internal}
