% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{compute_constant_copula_likelihood}
\alias{compute_constant_copula_likelihood}
\title{Compute Constant Copula Correlation Likelihood}
\usage{
compute_constant_copula_likelihood(
  residuals,
  weights,
  garch_pars,
  dist_pars,
  spec,
  transformation = "parametric",
  copula_dist = "mvn"
)
}
\arguments{
\item{residuals}{Matrix of residuals}

\item{weights}{Observation weights}

\item{garch_pars}{List of GARCH parameters}

\item{dist_pars}{Distribution parameters}

\item{spec}{Model specification}

\item{transformation}{PIT transformation type}

\item{copula_dist}{Copula distribution}
}
\value{
List with weighted_ll, dist_pars, warnings
}
\description{
Computes weighted log-likelihood for constant correlation copula
}
\keyword{internal}
