% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{compute_constant_correlation_likelihood}
\alias{compute_constant_correlation_likelihood}
\title{Compute Constant Correlation Weighted Likelihood}
\usage{
compute_constant_correlation_likelihood(
  residuals,
  weights,
  garch_pars,
  dist_pars,
  spec
)
}
\arguments{
\item{residuals}{residuals}

\item{weights}{weights}

\item{garch_pars}{garch_pars}

\item{dist_pars}{dist_pars}

\item{spec}{spec}
}
\description{
Compute Constant Correlation Weighted Likelihood
}
\keyword{internal}
