% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{compute_d2R_dQ2}
\alias{compute_d2R_dQ2}
\title{Compute d²R/dQ² (second derivative of correlation w.r.t. Q)}
\usage{
compute_d2R_dQ2(Q, dQ1, dQ2, d2Q, k)
}
\description{
Compute d²R/dQ² (second derivative of correlation w.r.t. Q)
}
\keyword{internal}
