% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockBootstrap.R
\name{compute_default_block_length}
\alias{compute_default_block_length}
\title{Compute Default Block Length Based on Autocorrelation (Internal Utility)}
\arguments{
\item{x}{A numeric matrix with rows as time points and columns as variables.}
}
\value{
An integer block length.
}
\description{
Computes a heuristic default block length for a multivariate time series
based on the average lag-1 autocorrelation.
}
\examples{
x <- matrix(rnorm(100), ncol = 2)
compute_default_block_length(x)

}
