% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_diagnostics.R
\name{compute_default_block_length_r}
\alias{compute_default_block_length_r}
\title{Compute Default Block Length (R version)}
\usage{
compute_default_block_length_r(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
Integer block length.
}
\description{
Computes a default block length based on average lag-1 autocorrelation.
}
\keyword{internal}
