% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{compute_dR_dQ}
\alias{compute_dR_dQ}
\title{Compute dR/dQ (derivative of correlation w.r.t. Q matrix)}
\usage{
compute_dR_dQ(Q, dQ, k)
}
\description{
Compute dR/dQ (derivative of correlation w.r.t. Q matrix)
}
\keyword{internal}
