% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{compute_mixing_angle}
\alias{compute_mixing_angle}
\title{Compute Mixing Matrix Angle}
\usage{
compute_mixing_angle(A_est, A_true)
}
\arguments{
\item{A_est}{Estimated mixing matrix}

\item{A_true}{True mixing matrix}
}
\value{
Scalar angle in radians
}
\description{
Computes the angle between estimated and true mixing matrices.
}
\keyword{internal}
