% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{compute_pit_transform}
\alias{compute_pit_transform}
\title{Compute Probability Integral Transform (PIT)}
\usage{
compute_pit_transform(
  std_residuals,
  uni_fit_list,
  transformation = "parametric",
  copula_dist = "mvn",
  dist_pars = NULL
)
}
\arguments{
\item{std_residuals}{Matrix of standardized residuals (T x k)}

\item{uni_fit_list}{List of fitted univariate GARCH models}

\item{transformation}{Type: "parametric", "empirical", or "spd"}

\item{copula_dist}{Copula distribution ("mvn" or "mvt")}

\item{dist_pars}{Distribution parameters}
}
\value{
Matrix of uniform margins (T x k)
}
\description{
Transforms standardized residuals to uniform [0,1] margins
}
\keyword{internal}
