% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{compute_spd_manual}
\alias{compute_spd_manual}
\title{Manual SPD Computation}
\usage{
compute_spd_manual(z, lower_threshold = 0.1, upper_threshold = 0.9)
}
\arguments{
\item{z}{Standardized residuals}

\item{lower_threshold}{Lower quantile threshold}

\item{upper_threshold}{Upper quantile threshold}
}
\value{
Vector of uniform values
}
\description{
Computes SPD transformation manually without tsdistributions.
Uses kernel density estimation for interior and empirical tails.
}
\keyword{internal}
