% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{compute_std_residuals}
\alias{compute_std_residuals}
\title{Compute Standardized Residuals from Simulated DCC-GARCH Data}
\usage{
compute_std_residuals(y, omega, alpha_garch, beta_garch)
}
\arguments{
\item{y}{T x k matrix of simulated returns}

\item{omega}{Vector of GARCH omega parameters (length k)}

\item{alpha_garch}{Vector of GARCH alpha parameters (length k)}

\item{beta_garch}{Vector of GARCH beta parameters (length k)}
}
\value{
T x k matrix of standardized residuals
}
\description{
Given simulated returns and known GARCH parameters, compute
the standardized residuals needed for DCC estimation.
}
