% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{copula_gradient}
\alias{copula_gradient}
\title{Copula Gradient for DCC(1,1) - Analytical}
\usage{
copula_gradient(
  params,
  z_matrix,
  weights,
  Qbar,
  copula_dist = "mvn",
  use_reparam = TRUE
)
}
\arguments{
\item{params}{Parameter vector: (psi, phi) if use_reparam=TRUE, (alpha, beta) otherwise.
For MVT copula, params\link{3} is the shape parameter.}

\item{z_matrix}{T x k matrix of copula residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance matrix}

\item{copula_dist}{"mvn" or "mvt"}

\item{use_reparam}{Logical; if TRUE, use reparameterized (psi, phi) space}
}
\value{
Gradient vector (same length as params)
}
\description{
Computes the analytical gradient of copula NLL w.r.t. (psi, phi)
or (alpha, beta) depending on use_reparam flag.
}
\keyword{internal}
