% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{copula_gradient_original}
\alias{copula_gradient_original}
\title{Copula Gradient in Original (alpha, beta) Space}
\usage{
copula_gradient_original(
  alpha,
  beta,
  z,
  weights,
  Qbar,
  copula_dist = "mvn",
  shape = NULL
)
}
\arguments{
\item{alpha}{DCC alpha parameter}

\item{beta}{DCC beta parameter}

\item{z}{T x k matrix of copula residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance}

\item{copula_dist}{"mvn" or "mvt"}

\item{shape}{Degrees of freedom (MVT only)}
}
\value{
Named vector c(alpha, beta)
}
\description{
Computes analytical gradient of copula NLL w.r.t. (alpha, beta)
}
\keyword{internal}
