% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{copula_gradient_shape}
\alias{copula_gradient_shape}
\title{Gradient of Copula NLL w.r.t. Shape (MVT)}
\usage{
copula_gradient_shape(shape, z, weights, alpha, beta, Qbar)
}
\arguments{
\item{shape}{Current degrees of freedom}

\item{z}{T x k matrix of copula residuals}

\item{weights}{Observation weights}

\item{alpha}{DCC alpha}

\item{beta}{DCC beta}

\item{Qbar}{Unconditional covariance}
}
\value{
Scalar gradient
}
\description{
Computes gradient of MVT copula NLL w.r.t. degrees of freedom.
Uses numerical differentiation for simplicity.
}
\keyword{internal}
