% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{copula_nll_mvn}
\alias{copula_nll_mvn}
\title{Copula NLL Wrapper (MVN)}
\usage{
copula_nll_mvn(params, z_matrix, weights, Qbar, copula = "mvn")
}
\description{
Wrapper for copula negative log-likelihood with MVN copula.
This references the function in tsbs_cgarch.R
}
\keyword{internal}
