% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{copula_nll_mvt}
\alias{copula_nll_mvt}
\title{Copula NLL for MVT}
\usage{
copula_nll_mvt(params, z_matrix, weights, Qbar)
}
\description{
Negative log-likelihood for Student-t copula
}
\keyword{internal}
