% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ms_varma_garch.R
\name{count_dcc_parameters}
\alias{count_dcc_parameters}
\title{Count DCC Parameters in Model}
\usage{
count_dcc_parameters(model_fits)
}
\description{
Count DCC Parameters in Model
}
\keyword{internal}
