% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{create_garch_spec_object_r}
\alias{create_garch_spec_object_r}
\title{Create a GARCH Specification Object (Convenience Function)}
\usage{
create_garch_spec_object_r(
  residuals,
  spec,
  model_type,
  current_pars,
  diagnostics = NULL,
  iteration = NULL,
  state = NULL,
  verbose = FALSE
)
}
\arguments{
\item{residuals}{Numeric}

\item{spec}{A spec list}

\item{model_type}{Character string}

\item{current_pars}{A list of parameters}

\item{diagnostics}{diagnostics}

\item{iteration}{iteration}

\item{state}{state}

\item{verbose}{verbose}
}
\value{
A GARCH specification object
}
\description{
Convenience function that handles the complex, model-specific
logic for creating both univariate and multivariate GARCH specification
objects for the purpose of log-likelihood calculation during the
EM-algorithm.
}
