% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{create_test_dcc_spec}
\alias{create_test_dcc_spec}
\title{Create a standard DCC spec for testing}
\usage{
create_test_dcc_spec(k = 2, distribution = "mvn")
}
\description{
Create a standard DCC spec for testing
}
