% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_analytical_hessian}
\alias{dcc11_analytical_hessian}
\title{Analytical Hessian of DCC(1,1) NLL}
\usage{
dcc11_analytical_hessian(
  params,
  std_resid,
  weights,
  Qbar,
  distribution = "mvn"
)
}
\arguments{
\item{params}{Parameter vector c(alpha, beta)}

\item{std_resid}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional correlation matrix}

\item{distribution}{"mvn" or "mvt"}
}
\value{
k x k Hessian matrix
}
\description{
Compute the analytical Hessian matrix of the DCC(1,1) negative
log-likelihood. This requires second derivatives through the DCC recursion.
}
\keyword{internal}
