% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{dcc11_gradient}
\alias{dcc11_gradient}
\title{Full DCC(1,1) Gradient Function for Optimizer}
\usage{
dcc11_gradient(
  params,
  std_resid,
  weights,
  Qbar,
  distribution = "mvn",
  use_reparam = TRUE
)
}
\arguments{
\item{params}{Named vector of parameters to optimize}

\item{std_resid}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance}

\item{distribution}{"mvn" or "mvt"}

\item{use_reparam}{Logical: use reparameterized (psi, phi) space?}
}
\value{
Vector of gradients (same length and order as params)
}
\description{
Compute gradient of weighted NLL w.r.t. all DCC parameters.
This is the main interface for use with optim().
}
