% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{dcc11_gradient_original}
\alias{dcc11_gradient_original}
\title{DCC(1,1) Weighted NLL Gradient (Original Parameterization)}
\usage{
dcc11_gradient_original(
  alpha,
  beta,
  z,
  weights,
  Qbar,
  distribution = "mvn",
  shape = NULL
)
}
\arguments{
\item{alpha}{DCC alpha parameter}

\item{beta}{DCC beta parameter}

\item{z}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance}

\item{distribution}{"mvn" or "mvt"}

\item{shape}{Degrees of freedom (MVT only)}
}
\value{
Named vector c(alpha, beta)
}
\description{
Compute gradient of weighted NLL w.r.t. (alpha, beta).
}
\keyword{internal}
