% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{dcc11_gradient_reparam}
\alias{dcc11_gradient_reparam}
\title{DCC(1,1) Weighted NLL Gradient (Reparameterized)}
\usage{
dcc11_gradient_reparam(
  psi,
  phi,
  z,
  weights,
  Qbar,
  distribution = "mvn",
  shape = NULL
)
}
\arguments{
\item{psi}{Unconstrained persistence parameter}

\item{phi}{Unconstrained ratio parameter}

\item{z}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance}

\item{distribution}{"mvn" or "mvt"}

\item{shape}{Degrees of freedom (MVT only)}
}
\value{
Named vector c(psi, phi)
}
\description{
Compute gradient of weighted NLL w.r.t. (psi, phi).
}
\keyword{internal}
