% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_hessian}
\alias{dcc11_hessian}
\title{Compute Hessian of DCC(1,1) NLL}
\usage{
dcc11_hessian(
  params,
  std_resid,
  weights,
  Qbar,
  distribution = "mvn",
  use_reparam = FALSE,
  method = "numerical",
  eps = 1e-05
)
}
\arguments{
\item{params}{MLE parameter estimates c(alpha, beta) or c(alpha, beta, shape)}

\item{std_resid}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional correlation matrix}

\item{distribution}{"mvn" or "mvt"}

\item{use_reparam}{Logical: parameters in reparameterized space?}

\item{method}{"numerical" (default) or "analytical"}

\item{eps}{Step size for numerical differentiation}
}
\value{
List with:
\item{hessian}{Hessian matrix of NLL}
\item{info}{Observed information matrix (= Hessian for NLL)}
\item{vcov}{Variance-covariance matrix (inverse of info)}
\item{se}{Standard errors}
\item{eigenvalues}{Eigenvalues of Hessian}
\item{condition_number}{Condition number of Hessian}
}
\description{
Compute the Hessian matrix of the negative log-likelihood.
This is a convenience wrapper that returns both the Hessian and derived
quantities (standard errors, eigenvalues).
}
