% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_nll_single_obs}
\alias{dcc11_nll_single_obs}
\title{NLL Contribution from Single Observation}
\usage{
dcc11_nll_single_obs(
  params,
  std_resid,
  weights,
  Qbar,
  t,
  distribution,
  use_reparam
)
}
\description{
Compute the NLL contribution from observation t.
}
\keyword{internal}
