% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{dcc11_recursion_with_grad}
\alias{dcc11_recursion_with_grad}
\title{DCC(1,1) Q-Matrix Recursion with Gradient Storage}
\usage{
dcc11_recursion_with_grad(z, alpha, beta, Qbar)
}
\arguments{
\item{z}{T x k matrix of standardized residuals}

\item{alpha}{DCC alpha parameter}

\item{beta}{DCC beta parameter}

\item{Qbar}{k x k unconditional covariance matrix}
}
\value{
List with Q, R, R_inv, log_det_R, dQ_dalpha, dQ_dbeta, success
}
\description{
Compute Q_t matrices and store gradient information for backpropagation.
}
\keyword{internal}
