% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{dcc11_reparam_jacobian}
\alias{dcc11_reparam_jacobian}
\title{Jacobian of Reparameterization}
\usage{
dcc11_reparam_jacobian(psi, phi)
}
\arguments{
\item{psi}{Unconstrained persistence parameter}

\item{phi}{Unconstrained ratio parameter}
}
\value{
2x2 matrix: rows = (alpha, beta), cols = (psi, phi)
}
\description{
Compute d(alpha, beta) / d(psi, phi) for chain rule.
}
\keyword{internal}
