% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_transform_se}
\alias{dcc11_transform_se}
\title{Transform Standard Errors Between Parameterizations}
\usage{
dcc11_transform_se(se_result, to_reparam = FALSE)
}
\arguments{
\item{se_result}{Result from dcc11_standard_errors()}

\item{to_reparam}{Logical: transform TO (psi, phi)? If FALSE, transform to (alpha, beta)}
}
\value{
Updated se_result with transformed standard errors
}
\description{
Transform standard errors from (psi, phi) to (alpha, beta) or
vice versa using the delta method.
}
