% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{dcc_gradient_shape}
\alias{dcc_gradient_shape}
\title{Gradient of NLL w.r.t. Shape Parameter (MVT)}
\usage{
dcc_gradient_shape(shape, z, weights, R_inv)
}
\arguments{
\item{shape}{Degrees of freedom (nu)}

\item{z}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{R_inv}{Array of inverse correlation matrices}
}
\value{
Scalar gradient (unnamed)
}
\description{
Compute gradient of weighted NLL w.r.t. degrees of freedom.

The MVT log-likelihood (per observation) is:
ll_t = lgamma((nu+k)/2) - lgamma(nu/2) - (k/2)\emph{log(pi}(nu-2))
- 0.5*log|R_t| - ((nu+k)/2)*log(kappa_t)

where kappa_t = 1 + q_t/(nu-2) and q_t = z_t' R_inv z_t.
}
\keyword{internal}
