% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ms_varma_garch.R
\name{detect_boundary_states}
\alias{detect_boundary_states}
\title{Detect States at DCC Parameter Boundary}
\usage{
detect_boundary_states(fit_result, threshold, criterion)
}
\description{
Detect States at DCC Parameter Boundary
}
\keyword{internal}
