% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{diagnostic_summary_stats}
\alias{diagnostic_summary_stats}
\title{Get Diagnostic Summary Statistics}
\usage{
diagnostic_summary_stats(diagnostics)
}
\arguments{
\item{diagnostics}{An object of class \code{ms_diagnostics}}
}
\value{
List with summary statistics
}
\description{
Compute summary statistics from diagnostic object.
}
