% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbs.R
\name{.estimate_p}
\alias{.estimate_p}
\title{Estimate p}
\usage{
.estimate_p(x, p_method, block_length, model_func, score_func)
}
\arguments{
\item{x}{x}

\item{p_method}{p estimation method}

\item{block_length}{Block length}

\item{model_func}{Model function for k fold cross validation}

\item{score_func}{Score function for k fold cross validation}
}
\value{
p
}
\description{
Estimate p
}
