% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbs.R
\name{.invalid_length}
\alias{.invalid_length}
\title{Invalid Length}
\usage{
.invalid_length(x, allow_null = TRUE)
}
\arguments{
\item{x}{x, vector, matrix or data.frame}

\item{allow_null}{NULL value may be allowed when functions calculates value
of x automatically.}
}
\value{
boolean
}
\description{
Invalid Length
}
\details{
As long as \code{exists()} is before the first "or", this will return TRUE,
when x doesn't exist.
}
