% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbs.R
\name{.invalid_percentage}
\alias{.invalid_percentage}
\title{Invalid percentage}
\usage{
.invalid_percentage(p, allow_null = TRUE)
}
\arguments{
\item{p}{Percentage (decimal fraction) to validate}

\item{allow_null}{Allow null?}
}
\value{
boolean
}
\description{
Invalid percentage
}
