% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbs.R
\name{.not_in_unit_interval}
\alias{.not_in_unit_interval}
\title{Value not in unit interval}
\usage{
.not_in_unit_interval(
  p,
  allow_null = TRUE,
  interval_type = c("open", "closed"),
  fail_mode = c("predictably", "gracefully")
)
}
\arguments{
\item{p}{Value to validate, e.g. a percentage (decimal fraction) .}

\item{allow_null}{boolean, allow NULL?}

\item{interval_type}{One of \code{"open"} or \code{"closed"}. Indicates if the valid
interval is an open or closed unit interval, i.e. \eqn{p \in (0, 1)} or
\eqn{p \in [0, 1]} respectively.}

\item{fail_mode}{How to handle validation errors: "predictably" (fail fast)
or "gracefully" (return FALSE on error)}
}
\value{
boolean
}
\description{
Returns TRUE if \code{p} is not in the unit interval, FALSE otherwise.
}
\details{
If
}
