% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{.sample_blocks}
\alias{.sample_blocks}
\title{Sample bootstrap blocks}
\usage{
.sample_blocks(
  x,
  n_boot,
  num_blocks,
  states,
  num_boots,
  parallel = FALSE,
  num_cores = 1L
)
}
\arguments{
\item{x}{x}

\item{n_boot}{Length of bootstrap series}

\item{num_blocks}{Number of blocks}

\item{states}{States (see tsbs() help)}

\item{num_boots}{Number of bootstraps}

\item{parallel}{Boolean}

\item{num_cores}{Number of cores}
}
\description{
Deprecated. Use \code{.sample_blocks_mv()} instead.
}
\details{
Internal helper function.
}
