% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{dt_scaled_log}
\alias{dt_scaled_log}
\title{Log-density of Scaled Student-t}
\usage{
dt_scaled_log(x, shape)
}
\arguments{
\item{x}{Value}

\item{shape}{Degrees of freedom}
}
\value{
Log-density
}
\description{
Computes log-density of standardized Student-t (mean 0, var 1)
}
\keyword{internal}
