% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_arma_weighted_r}
\alias{estimate_arma_weighted_r}
\title{Estimate Conditional Mean Parameters (R Helper)}
\usage{
estimate_arma_weighted_r(y, weights, spec, model_type = "univariate")
}
\arguments{
\item{y}{}

\item{weights}{}

\item{spec}{}

\item{model_type}{}
}
\description{
Estimate Conditional Mean Parameters (R Helper)
}
