% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_inference.R
\name{estimate_component_garch}
\alias{estimate_component_garch}
\title{Estimate Component GARCH}
\usage{
estimate_component_garch(component, weights, start_pars, distribution = "norm")
}
\description{
Estimate GARCH(1,1) parameters for a single component.
}
\keyword{internal}
