% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc.R
\name{estimate_dcc11_reparam}
\alias{estimate_dcc11_reparam}
\title{Estimate DCC(1,1) using reparameterized optimization}
\usage{
estimate_dcc11_reparam(
  std_resid,
  weights,
  dcc_start_pars,
  Qbar,
  distribution = "mvn",
  dist_pars = NULL,
  verbose = FALSE
)
}
\description{
Estimate DCC(1,1) using reparameterized optimization
}
\keyword{internal}
