% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_dcc_parameters_weighted}
\alias{estimate_dcc_parameters_weighted}
\title{Estimate DCC Correlation Dynamics (Weighted)}
\usage{
estimate_dcc_parameters_weighted(
  residuals,
  weights,
  garch_pars,
  dcc_start_pars,
  dist_start_pars,
  spec,
  diagnostics = NULL,
  iteration = NULL,
  state = NULL,
  verbose = FALSE
)
}
\description{
Estimate DCC Correlation Dynamics (Weighted)
}
\keyword{internal}
