% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_dcc_pq_penalty}
\alias{estimate_dcc_pq_penalty}
\title{Estimate DCC(p,q) using penalty-based optimization}
\usage{
estimate_dcc_pq_penalty(
  std_resid,
  weights,
  dcc_start_pars,
  Qbar,
  distribution = "mvn",
  dist_pars = NULL,
  verbose = FALSE
)
}
\description{
Estimate DCC(p,q) using penalty-based optimization
}
\keyword{internal}
