% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_garch_weighted_copula}
\alias{estimate_garch_weighted_copula}
\title{Copula Weighted Estimation (Placeholder)}
\usage{
estimate_garch_weighted_copula(residuals, weights, spec)
}
\description{
Copula Weighted Estimation (Placeholder)
}
\keyword{internal}
