% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_garch_weighted_dcc}
\alias{estimate_garch_weighted_dcc}
\title{DCC Weighted Estimation}
\usage{
estimate_garch_weighted_dcc(
  residuals,
  weights,
  spec,
  diagnostics = NULL,
  iteration = NULL,
  state = NULL,
  verbose = FALSE
)
}
\description{
DCC Weighted Estimation
}
\keyword{internal}
