% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_garch_weighted_gogarch}
\alias{estimate_garch_weighted_gogarch}
\title{GOGARCH Weighted Estimation (Placeholder)}
\usage{
estimate_garch_weighted_gogarch(residuals, weights, spec)
}
\description{
GOGARCH Weighted Estimation (Placeholder)
}
\keyword{internal}
