% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_garch_weighted_multivariate}
\alias{estimate_garch_weighted_multivariate}
\title{Multivariate GARCH Parameter Estimation (Two-Stage)}
\usage{
estimate_garch_weighted_multivariate(
  residuals,
  weights,
  spec,
  diagnostics = NULL,
  iteration = NULL,
  state = NULL,
  verbose = FALSE,
  dcc_threshold = 0.02,
  dcc_criterion = "bic",
  force_constant = FALSE
)
}
\arguments{
\item{residuals}{residuals}

\item{weights}{weights}

\item{spec}{spec}

\item{diagnostics}{diagnostics}

\item{iteration}{iteration}

\item{state}{state}

\item{verbose}{verbose}

\item{dcc_threshold}{dcc_threshold}

\item{dcc_criterion}{dcc_criterion}

\item{force_constant}{force_constant}
}
\description{
Implements weighted MLE for multivariate GARCH models
Stage 1: Estimate univariate GARCH parameters for each series
Stage 2: Estimate dependence parameters (DCC/Copula) or GOGARCH rotation
}
\keyword{internal}
