% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_garch_weighted_univariate}
\alias{estimate_garch_weighted_univariate}
\title{Univariate GARCH + Distribution Parameter Estimation}
\usage{
estimate_garch_weighted_univariate(residuals, weights, spec, verbose)
}
\description{
Univariate GARCH + Distribution Parameter Estimation
}
\keyword{internal}
