% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogarch_ica_improvements.R
\name{evaluate_ica_quality}
\alias{evaluate_ica_quality}
\title{Evaluate ICA Quality}
\usage{
evaluate_ica_quality(S)
}
\arguments{
\item{S}{T x n_components matrix of independent components}
}
\value{
List with:
\describe{
\item{independence_score}{1 - mean(|cor(S)|), higher is better}
\item{total_negentropy}{Sum of component negentropies}
\item{component_kurtosis}{Excess kurtosis per component}
\item{correlation_matrix}{Component correlation matrix}
}
}
\description{
Computes quality metrics for ICA decomposition.
}
\seealso{
\code{\link{improved_ica_decomposition}}, \code{\link{gogarch_diagnostics}}
}
