% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_diagnostics.R
\name{extract_summary_stats}
\alias{extract_summary_stats}
\title{Extract Summary Statistics from Diagnostics}
\usage{
extract_summary_stats(diagnostics)
}
\arguments{
\item{diagnostics}{A \code{tsbs_diagnostics} object.}
}
\value{
A list with summary statistics.
}
\description{
Extract Summary Statistics from Diagnostics
}
