% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{fit_gogarch_model}
\alias{fit_gogarch_model}
\title{Fit GOGARCH Model}
\usage{
fit_gogarch_model(y, k, ica_method = "radical", distribution = "norm")
}
\arguments{
\item{y}{Matrix of returns (T x k)}

\item{k}{Number of components}

\item{ica_method}{ICA algorithm ("radical" or "fastica")}

\item{distribution}{Component distribution}
}
\value{
List with fitted parameters and convergence info
}
\description{
Fits a GOGARCH model to multivariate data.
}
\keyword{internal}
