% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_ms_varma_garch_cpp}
\alias{fit_ms_varma_garch_cpp}
\title{Fit a general MS-ARMA-GARCH model via the EM Algorithm in C++}
\usage{
fit_ms_varma_garch_cpp(y, M, spec, model_type, control)
}
\arguments{
\item{y}{A (T x k) matrix of (differenced) time series data.}

\item{M}{The number of states.}

\item{spec}{A list of model specifications from R.}

\item{model_type}{"univariate" or "multivariate".}

\item{control}{A list with max_iter and tol.}
}
\value{
A raw list with estimated parameters and results.
}
\description{
Internal C++ function to orchestrate the EM estimation.
Delegates all statistical calculations (likelihood,
parameter estimation) to helper functions in R to interface
with TMB-based packages like tsgarch.
}
