% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{fit_univariate_garch}
\alias{fit_univariate_garch}
\title{Fit Univariate GARCH(1,1)}
\usage{
fit_univariate_garch(y, distribution = "norm")
}
\description{
Fits a simple GARCH(1,1) model to a univariate series.
}
\keyword{internal}
