% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{generate_tsmarch_parnames}
\alias{generate_tsmarch_parnames}
\title{Generate Correct Parameter Names for tsmarch}
\usage{
generate_tsmarch_parnames(pars)
}
\arguments{
\item{pars}{A nested parameter list.}
}
\value{
A flat named list.
}
\description{
Translates a nested parameter list into the flat named list
that tsmarch's parmatrix expects (e.g., "omega\link{1}").
}
\keyword{internal}
