% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-utils.R
\name{get_demo_path}
\alias{get_demo_path}
\title{Get path to a specific demo Rmd file}
\usage{
get_demo_path(name)
}
\arguments{
\item{name}{Name of the demo file (with or without .Rmd extension)}
}
\value{
Full path to the demo Rmd file
}
\description{
Get path to a specific demo Rmd file
}
\examples{
# Get path to a demo
demo_path <- get_demo_path("example-demo")

# If you want to render it
# rmarkdown::render(demo_path, output_dir = tempdir())
}
