% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{grad_nll_wrt_R_copula_mvn}
\alias{grad_nll_wrt_R_copula_mvn}
\title{Gradient of Copula NLL w.r.t. R_t (MVN)}
\usage{
grad_nll_wrt_R_copula_mvn(z_t, R_t, R_inv_t)
}
\arguments{
\item{z_t}{k-vector of copula residuals at time t}

\item{R_t}{k x k correlation matrix}

\item{R_inv_t}{k x k inverse correlation matrix}
}
\value{
k x k gradient matrix
}
\description{
Compute gradient of Gaussian copula NLL contribution w.r.t. R_t.
For Gaussian copula: LL_t = -0.5 * (log|R| + z'R^{-1}z - z'z)
Gradient: d(-LL_t)/dR = 0.5 * (R^{-1} - R^{-1}zz'R^{-1})
}
\keyword{internal}
