% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{grad_nll_wrt_R_copula_mvt}
\alias{grad_nll_wrt_R_copula_mvt}
\title{Gradient of Copula NLL w.r.t. R_t (MVT)}
\usage{
grad_nll_wrt_R_copula_mvt(z_t, R_t, R_inv_t, shape)
}
\arguments{
\item{z_t}{k-vector of copula residuals at time t}

\item{R_t}{k x k correlation matrix}

\item{R_inv_t}{k x k inverse correlation matrix}

\item{shape}{Degrees of freedom}
}
\value{
k x k gradient matrix
}
\description{
Compute gradient of Student-t copula NLL contribution w.r.t. R_t.
For Student-t copula, the gradient has an additional scaling factor.
}
\keyword{internal}
