% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{grad_nll_wrt_R_mvt}
\alias{grad_nll_wrt_R_mvt}
\title{Gradient of NLL w.r.t. R_t (MVT)}
\usage{
grad_nll_wrt_R_mvt(z_t, R_t, R_inv_t, shape)
}
\arguments{
\item{z_t}{k-vector of standardized residuals at time t}

\item{R_t}{k x k correlation matrix}

\item{R_inv_t}{k x k inverse correlation matrix}

\item{shape}{Degrees of freedom}
}
\value{
k x k gradient matrix
}
\description{
Compute gradient of negative log-likelihood contribution w.r.t. R_t.
}
\keyword{internal}
