% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_gradient.R
\name{grad_R_to_Q}
\alias{grad_R_to_Q}
\title{Gradient of NLL w.r.t. Q_t (via R_t normalization)}
\usage{
grad_R_to_Q(grad_R, Q_t, R_t)
}
\arguments{
\item{grad_R}{k x k gradient w.r.t. R_t}

\item{Q_t}{k x k Q matrix}

\item{R_t}{k x k correlation matrix}
}
\value{
k x k gradient w.r.t. Q_t
}
\description{
Backpropagate gradient from R_t to Q_t through the normalization.
}
\keyword{internal}
