% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{grad_R_to_Q_copula}
\alias{grad_R_to_Q_copula}
\title{Gradient of NLL w.r.t. Q_t (Copula)}
\usage{
grad_R_to_Q_copula(grad_R, Q_t, R_t)
}
\arguments{
\item{grad_R}{k x k gradient w.r.t. R_t}

\item{Q_t}{k x k Q matrix}

\item{R_t}{k x k correlation matrix}
}
\value{
k x k gradient w.r.t. Q_t
}
\description{
Backpropagate gradient from R_t to Q_t through the normalization.
R_t = D^{-1} Q_t D^{-1} where D = diag(sqrt(diag(Q_t)))
}
\keyword{internal}
