% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{is_dcc_11}
\alias{is_dcc_11}
\title{Check if DCC specification is DCC(1,1)}
\usage{
is_dcc_11(dcc_params)
}
\arguments{
\item{dcc_params}{Named list of DCC parameters}
}
\value{
Logical; TRUE if this is a DCC(1,1) specification
}
\description{
Check if DCC specification is DCC(1,1)
}
\keyword{internal}
