% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-utils.R
\name{list_package_demos}
\alias{list_package_demos}
\title{List available demonstration Rmd files}
\usage{
list_package_demos()
}
\value{
Character vector of available demo Rmd file names
}
\description{
These demonstrations are computationally intensive and not built as vignettes.
Use \code{get_demo_path()} to get the full path to a specific demo.
}
\examples{
list_package_demos()
}
