% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error (MAE)}
\usage{
mae(pred, actual)
}
\arguments{
\item{actual}{Actual values.}

\item{predicted}{Predicted values.}
}
\value{
MAE
}
\description{
Mean Absolute Error (MAE)
}
