% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{ms_varma_garch_bs}
\alias{ms_varma_garch_bs}
\title{Stationary Bootstrap for a General MS-VARMA-GARCH Model}
\usage{
ms_varma_garch_bs(
  x,
  n_boot = NULL,
  num_blocks = 100,
  num_boots = 100,
  M,
  d = 0,
  spec,
  model_type = c("univariate", "multivariate"),
  control = list(),
  parallel = FALSE,
  num_cores = 1L
)
}
\arguments{
\item{x}{A numeric matrix or data frame where rows are observations and
columns are the time series variables.}

\item{n_boot}{An integer specifying the length of each bootstrapped series.
If NULL (the default), the length of the original series is used.}

\item{num_blocks}{An integer specifying the number of blocks to sample for
the bootstrap. Defaults to 100.}

\item{num_boots}{An integer specifying the total number of bootstrap samples
to generate. Defaults to 100.}

\item{M}{An integer specifying the number of states in the Markov chain.}

\item{d}{An integer specifying the order of differencing for the ARIMA model.}

\item{spec}{A list of model specifications, one for each of the M states.}

\item{model_type}{A character string, either "univariate" or "multivariate".}

\item{control}{A list of control parameters for the EM algorithm.}

\item{parallel}{A logical value indicating whether to use parallel processing.}

\item{num_cores}{An integer specifying the number of cores for parallel processing.}
}
\value{
A list of bootstrapped time series matrices.
}
\description{
This function fits a flexible N-state Markov-Switching model and then uses
the estimated state sequence to perform a stationary block bootstrap. This
generates resampled time series that preserve the state-dependent properties
of the original data.
}
