% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msar_bootstrap.R
\name{msar_bootstrap}
\alias{msar_bootstrap}
\title{Markov Switching Autoregressive (MSAR) Bootstrap for Time Series}
\usage{
msar_bootstrap(
  returns,
  ar_order = 1,
  nstates = 2,
  num_blocks = 100,
  num_boots = 100,
  shuffle_blocks = TRUE
)
}
\arguments{
\item{num_boots}{Number of bootstrap samples to generate.}

\item{x}{A numeric matrix with one column (univariate time series).}

\item{k}{Number of states in the MSAR model.}

\item{order}{AR order of each regime.}
}
\value{
A list of bootstrap replicates (numeric vectors).
}
\description{
Fits an MSAR model and generates bootstrap replicates.
}
\examples{
set.seed(456)
returns <- arima.sim(n = 200, list(ar = 0.7))
boot_samples <- msar_bootstrap(returns, ar_order = 1, nstates = 2, num_blocks = 10, num_boots = 50)

}
