% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{mse}
\alias{mse}
\title{Mean Squared Error (MSE)}
\usage{
mse(pred, actual)
}
\arguments{
\item{actual}{Actual values.}

\item{predicted}{Predicted values.}
}
\value{
MSE
}
\description{
Mean Squared Error (MSE)
}
