% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-utils.R
\name{open_demo}
\alias{open_demo}
\title{Open a demo Rmd file in RStudio or default editor}
\usage{
open_demo(name)
}
\arguments{
\item{name}{Name of the demo file (with or without .Rmd extension)}
}
\description{
Open a demo Rmd file in RStudio or default editor
}
