% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{perform_m_step_parallel_r_old}
\alias{perform_m_step_parallel_r_old}
\title{Perform the M-Step in Parallel (R Helper)}
\usage{
perform_m_step_parallel_r_old(y, weights, spec, model_type)
}
\arguments{
\item{y}{The time series data.}

\item{weights}{The (T x M) matrix of smoothed probabilities from the E-step.}

\item{spec}{The full list of model specifications.}

\item{model_type}{"univariate" or "multivariate".}
}
\value{
A list of length M containing the updated model fits for each state.
}
\description{
This function is called once per EM iteration from C++. It uses
the 'future' framework to estimate the parameters for all M states in parallel.
}
