% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.ms_diagnostics}
\alias{plot.ms_diagnostics}
\title{Plot Diagnostics}
\usage{
\method{plot}{ms_diagnostics}(x, type = c("all", "ll_evolution", "parameters", "sigma"), ...)
}
\description{
Creates diagnostic plots for EM convergence
}
