% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{plot_cgarch_nll_contours}
\alias{plot_cgarch_nll_contours}
\title{Plot CGARCH NLL Contours}
\usage{
plot_cgarch_nll_contours(
  surface,
  true_params = NULL,
  mle_params = NULL,
  title = NULL
)
}
\arguments{
\item{surface}{Result from compute_cgarch_nll_surface()}

\item{true_params}{Optional vector c(alpha, beta) of true parameters}

\item{mle_params}{Optional vector c(alpha, beta) of MLE estimates}

\item{title}{Plot title}
}
\value{
plotly object
}
\description{
Create an interactive contour plot of the CGARCH likelihood surface.
}
