% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{plot_coverage_diagnostic}
\alias{plot_coverage_diagnostic}
\title{Plot Coverage Diagnostic}
\usage{
plot_coverage_diagnostic(mc_result, param = "alpha", max_show = 50)
}
\arguments{
\item{mc_result}{Result from run_dcc_monte_carlo()}

\item{param}{Which parameter ("alpha" or "beta")}

\item{max_show}{Maximum number of replications to show}
}
\value{
plotly object
}
\description{
Visualize confidence interval coverage for each replication.
}
