% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{plot_mc_estimates}
\alias{plot_mc_estimates}
\title{Plot Monte Carlo Estimate Distribution}
\usage{
plot_mc_estimates(mc_result)
}
\arguments{
\item{mc_result}{Result from run_dcc_monte_carlo()}
}
\value{
plotly object with subplots
}
\description{
Create histograms and scatter plots of MC estimates.
}
