% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{plot_mixing_matrix}
\alias{plot_mixing_matrix}
\title{Plot GOGARCH Mixing Matrix Heatmap}
\usage{
plot_mixing_matrix(A, title = "GOGARCH Mixing Matrix")
}
\arguments{
\item{A}{Mixing matrix (k x k)}

\item{title}{Plot title}
}
\value{
plotly object with heatmap
}
\description{
Visualize the estimated mixing matrix from GOGARCH.
}
