% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_diagnostics.R
\name{plot_nll_surface_3d}
\alias{plot_nll_surface_3d}
\title{Plot 3D NLL Surface with plotly}
\usage{
plot_nll_surface_3d(
  surface,
  true_params = NULL,
  mle_params = NULL,
  title = "DCC(1,1) Negative Log-Likelihood Surface"
)
}
\arguments{
\item{surface}{Result from compute_nll_surface()}

\item{true_params}{Optional vector c(alpha, beta) of true parameters}

\item{mle_params}{Optional vector c(alpha, beta) of MLE estimates}

\item{title}{Plot title}
}
\value{
plotly object
}
\description{
Create an interactive 3D surface plot of the DCC likelihood.
}
