% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_inference.R
\name{plot_profile_likelihood}
\alias{plot_profile_likelihood}
\title{Plot Profile Likelihood}
\usage{
plot_profile_likelihood(profile_result, param_name = "alpha")
}
\arguments{
\item{profile_result}{Result from profile_one_parameter or dcc11_profile_likelihood_ci}

\item{param_name}{Which parameter to plot ("alpha" or "beta")}
}
\value{
plotly object
}
\description{
Create a visualization of the profile likelihood with CI.
}
