% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_diagnostics.R
\name{print.tsbs_benchmark}
\alias{print.tsbs_benchmark}
\title{Print Method for tsbs_benchmark}
\usage{
\method{print}{tsbs_benchmark}(x, ...)
}
\arguments{
\item{x}{A \code{tsbs_benchmark} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for tsbs_benchmark
}
