% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_diagnostics.R
\name{print.tsbs_diagnostics}
\alias{print.tsbs_diagnostics}
\title{Print Method for Bootstrap Diagnostics}
\usage{
\method{print}{tsbs_diagnostics}(x, ...)
}
\arguments{
\item{x}{A \code{tsbs_diagnostics} object.}

\item{...}{Additional arguments passed to summary.}
}
\value{
Invisibly returns the object.
}
\description{
Print Method for Bootstrap Diagnostics
}
