% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_inference.R
\name{print_inference_comparison}
\alias{print_inference_comparison}
\title{Print Inference Comparison}
\usage{
print_inference_comparison(inf_result)
}
\arguments{
\item{inf_result}{Result from dcc11_comprehensive_inference}
}
\description{
Print a formatted comparison of inference methods.
}
