% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_inference.R
\name{process_bootstrap_results}
\alias{process_bootstrap_results}
\title{Process Bootstrap Results}
\usage{
process_bootstrap_results(
  boot_estimates,
  mle_params,
  param_names,
  n_boot,
  verbose = TRUE
)
}
\description{
Compute bootstrap statistics from replications.
}
\keyword{internal}
