% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_diagnostics.R
\name{record_original_stats}
\alias{record_original_stats}
\title{Record Original Series Statistics}
\usage{
record_original_stats(diagnostics, x)
}
\arguments{
\item{diagnostics}{A \code{tsbs_diagnostics} object.}

\item{x}{Original data matrix.}
}
\value{
Updated diagnostics object.
}
\description{
Record Original Series Statistics
}
\keyword{internal}
