% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_diagnostics.R
\name{run_complete_gogarch_diagnostics}
\alias{run_complete_gogarch_diagnostics}
\title{Complete GOGARCH Diagnostics}
\usage{
run_complete_gogarch_diagnostics(
  n_sim = 50,
  n_obs = 500,
  k = 3,
  omega = NULL,
  alpha_garch = NULL,
  beta_garch = NULL,
  distribution = "norm",
  ica_method = "radical",
  seed = 42,
  plot = TRUE
)
}
\arguments{
\item{n_sim}{Number of MC replications}

\item{n_obs}{Observations per replication}

\item{k}{Number of series/components}

\item{omega}{Vector of component GARCH omega}

\item{alpha_garch}{Vector of component GARCH alpha}

\item{beta_garch}{Vector of component GARCH beta}

\item{distribution}{Component distribution}

\item{ica_method}{ICA algorithm}

\item{seed}{Random seed}

\item{plot}{Logical: create plots?}
}
\value{
List with diagnostic results
}
\description{
Run comprehensive diagnostics for GOGARCH models.
}
