% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_se.R
\name{simulate_adcc_residuals}
\alias{simulate_adcc_residuals}
\title{Simulate ADCC Residuals}
\usage{
simulate_adcc_residuals(
  n,
  k,
  alpha,
  gamma,
  beta,
  Qbar,
  Nbar,
  copula_dist = "mvn",
  shape = NULL
)
}
\description{
Simulate residuals from an ADCC model for parametric bootstrap.
}
\keyword{internal}
