% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_inference.R
\name{simulate_copula_residuals}
\alias{simulate_copula_residuals}
\title{Simulate Copula Residuals}
\usage{
simulate_copula_residuals(
  n,
  k,
  alpha,
  beta,
  Qbar,
  copula_dist = "mvn",
  shape = NULL
)
}
\description{
Simulate residuals from a DCC-style copula for parametric bootstrap.
}
\keyword{internal}
