% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_inference.R
\name{simulate_dcc_residuals}
\alias{simulate_dcc_residuals}
\title{Simulate DCC Standardized Residuals}
\usage{
simulate_dcc_residuals(
  n,
  k,
  alpha,
  beta,
  Qbar,
  distribution = "mvn",
  shape = NULL
)
}
\description{
Simulate standardized residuals from a DCC(1,1) model.
}
\keyword{internal}
