% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_inference.R
\name{simulate_garch_component}
\alias{simulate_garch_component}
\title{Simulate GARCH Component}
\usage{
simulate_garch_component(n, pars, distribution = "norm")
}
\description{
Simulate a univariate GARCH(1,1) series for parametric bootstrap.
}
\keyword{internal}
