% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbs-package.R
\docType{package}
\name{tsbs-package}
\alias{tsbs-package}
\title{tsbs: Flexible Block Bootstrap for Time Series}
\description{
Provides advanced block bootstrap algorithms for time series data, including moving, stationary, wild, HMM-based, and MSAR-based bootstraps. Includes automatic parameter tuning and high-performance C++ routines.
}
\author{
\strong{Maintainer}: Martin Hoshi Vognsen \email{martinvognsen@gmail.com} (\href{https://orcid.org/0000-0003-4460-7763}{ORCID})

}
\keyword{internal}
