% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_bootstrap.R
\name{wild_bootstrap}
\alias{wild_bootstrap}
\title{Wild Bootstrap for Residual Resampling}
\usage{
wild_bootstrap(x, num_boots)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{num_boots}{Number of bootstrap replicates.}
}
\value{
A list of bootstrapped samples.
}
\description{
Performs wild bootstrap by multiplying residuals by random weights (+1 or -1).
}
\examples{
set.seed(789)
resids <- rnorm(100)
boot_samples <- wild_bootstrap(resids, num_boots = 50)

}
