% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{calculate_loglik_vector_r}
\alias{calculate_loglik_vector_r}
\title{Calculate the Log-Likelihood Vector (R Helper)}
\usage{
calculate_loglik_vector_r(
  y,
  current_pars,
  spec,
  model_type = "univariate",
  diagnostics = NULL,
  iteration = NULL,
  state = NULL,
  verbose = FALSE
)
}
\arguments{
\item{y}{}

\item{current_pars}{}

\item{spec}{}

\item{model_type}{}
}
\description{
Calculate the Log-Likelihood Vector (R Helper)
}
