% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{compute_copula_residuals}
\alias{compute_copula_residuals}
\title{Compute Copula Residuals from Uniform Margins}
\usage{
compute_copula_residuals(u_matrix, copula_dist = "mvn", dist_pars = NULL)
}
\arguments{
\item{u_matrix}{Matrix of uniform margins (T x k)}

\item{copula_dist}{Copula distribution ("mvn" or "mvt")}

\item{dist_pars}{Distribution parameters (shape for mvt)}
}
\value{
Matrix of copula residuals (T x k)
}
\description{
Transforms uniform margins to copula residuals (Z) based on
the copula distribution
}
\keyword{internal}
