% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgarch.R
\name{copula_nll}
\alias{copula_nll}
\title{Copula Negative Log-Likelihood for DCC(1,1)}
\usage{
copula_nll(
  params,
  z_matrix,
  weights,
  Qbar,
  copula_dist = "mvn",
  use_reparam = TRUE
)
}
\arguments{
\item{params}{Parameter vector (psi, phi) or (psi, phi, shape)}

\item{z_matrix}{Matrix of copula residuals (T x k)}

\item{weights}{Observation weights}

\item{Qbar}{Unconditional covariance matrix}

\item{copula_dist}{Copula distribution ("mvn" or "mvt")}

\item{use_reparam}{Logical; if TRUE, params are in reparameterized space}
}
\value{
Scalar negative log-likelihood
}
\description{
Computes the weighted copula negative log-likelihood
}
\keyword{internal}
