% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{create_garch_spec_object_r}
\alias{create_garch_spec_object_r}
\title{Create a GARCH Specification Object (Convenience Function)}
\usage{
create_garch_spec_object_r(residuals, spec, model_type)
}
\description{
This is a convenience function that handles the complex,
model-specific logic for creating both univariate and multivariate GARCH
specification objects.
}
