% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_analytical_hessian}
\alias{dcc11_analytical_hessian}
\title{Analytical Hessian of DCC(1,1) Negative Log-Likelihood}
\usage{
dcc11_analytical_hessian(
  params,
  std_resid,
  weights,
  Qbar,
  distribution = "mvn"
)
}
\arguments{
\item{params}{Parameter vector c(alpha, beta)}

\item{std_resid}{T x k matrix of standardized residuals from first-stage
GARCH}

\item{weights}{T-vector of observation weights (typically all 1s)}

\item{Qbar}{k x k unconditional correlation matrix (sample correlation of
std_resid)}

\item{distribution}{"mvn" (multivariate normal) - mvt not yet supported
analytically}
}
\value{
2 x 2 Hessian matrix of the NLL with respect to (alpha, beta)
}
\description{
Computes the exact Hessian matrix of the DCC(1,1) negative
log-likelihood using analytical derivatives. This is more accurate than
numerical differentiation but only implemented for the 2-parameter (alpha,
beta) case.
}
\details{
The DCC(1,1) model specifies: \deqn{Q_t = (1 - \alpha - \beta)
\bar{Q} + \alpha z_{t-1} z_{t-1}' + \beta Q_{t-1}} \deqn{R_t =
diag(Q_t)^{-1/2} Q_t diag(Q_t)^{-1/2}}

The correlation log-likelihood contribution at time t is: \deqn{\ell_t =
-\frac{1}{2}[\log|R_t| + z_t' R_t^{-1} z_t - z_t' z_t]}

This function computes \eqn{\partial^2 (-\sum_t \ell_t) / \partial \theta
\partial \theta'} where \eqn{\theta = (\alpha, \beta)'}.
}
\keyword{internal}
