% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_sandwich_se}
\alias{dcc11_sandwich_se}
\title{Compute Sandwich (Robust) Standard Errors}
\usage{
dcc11_sandwich_se(
  params,
  std_resid,
  weights,
  Qbar,
  distribution = "mvn",
  use_reparam = FALSE
)
}
\arguments{
\item{params}{MLE parameter estimates}

\item{std_resid}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance matrix}

\item{distribution}{"mvn" or "mvt"}

\item{use_reparam}{Logical: parameters in (psi, phi) space?}
}
\value{
List with:
\item{se}{Standard errors}
\item{vcov}{Variance-covariance matrix (inverse of info)}
\item{bread}{Bread}
\item{meatinfo}{Meat}
\item{vcov}{Variance-covariance matrix (inverse of info)}
\item{params}{Parameters}
\item{param_names}{Parameter names}
\item{method}{"sandwich"}
}
\description{
Heteroskedasticity-robust SEs using sandwich estimator:
Var(theta) = I^{-1} J I^{-1}
}
