% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_standard_errors}
\alias{dcc11_standard_errors}
\title{Compute Standard Errors for DCC(1,1) Parameters}
\usage{
dcc11_standard_errors(
  params,
  std_resid,
  weights,
  Qbar,
  distribution = "mvn",
  use_reparam = FALSE,
  method = c("hessian", "sandwich")
)
}
\arguments{
\item{params}{MLE parameter estimates}

\item{std_resid}{T x k matrix of standardized residuals}

\item{weights}{T-vector of observation weights}

\item{Qbar}{k x k unconditional covariance matrix}

\item{distribution}{"mvn" or "mvt"}

\item{use_reparam}{Logical: parameters in (psi, phi) space?}

\item{method}{SE method: "hessian" (default) or "sandwich"}
}
\value{
List with:
If \code{method="hessian"}:
\item{hessian}{Hessian matrix of NLL}
\item{info}{Observed information matrix (= Hessian for NLL)}
\item{vcov}{Variance-covariance matrix (inverse of info)}
\item{se}{Standard errors}
\item{eigenvalues}{Eigenvalues of Hessian}
\item{eigenvectors}{Eigenvectors of Hessian}
\item{condition_number}{Condition number of Hessian}
\item{params}{Parameters}
\item{param_names}{Para,eter names}
\item{method}{"hessian"}

If \code{method="sandwich"}:
\item{se}{Standard errors}
\item{vcov}{Variance-covariance matrix (inverse of info)}
\item{bread}{Bread}
\item{meatinfo}{Meat}
\item{vcov}{Variance-covariance matrix (inverse of info)}
\item{params}{Parameters}
\item{param_names}{Parameter names}
\item{method}{"sandwich"}
}
\description{
Main user-facing function for SE computation. Dispatches to
appropriate method based on the \code{method} argument.
}
\details{
For reliable beta inference in high-persistence DCC models, consider using
bootstrap SEs or profile likelihood CIs from \code{dcc_inference.R}.
}
