% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_hessian.R
\name{dcc11_transform_se}
\alias{dcc11_transform_se}
\title{Transform Standard Errors Between Parameterizations}
\usage{
dcc11_transform_se(se_result, to_reparam = FALSE)
}
\arguments{
\item{se_result}{Result from dcc11_standard_errors() or dcc11_hessian()}

\item{to_reparam}{Logical: transform TO (psi, phi)?}
}
\value{
Updated se_result with transformed SEs
}
\description{
Transform Standard Errors Between Parameterizations
}
