% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{.sample_blocks_mv}
\alias{.sample_blocks_mv}
\title{Helper function to sample blocks from a multivariate series}
\usage{
.sample_blocks_mv(
  x,
  n_boot,
  num_blocks,
  states,
  num_boots,
  parallel = FALSE,
  num_cores = 1L
)
}
\description{
Helper function to sample blocks from a multivariate series
}
