% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-varma-garch_helper_functions.R
\name{estimate_garch_weighted_r}
\alias{estimate_garch_weighted_r}
\title{Estimate GARCH Parameters (R Helper)}
\usage{
estimate_garch_weighted_r(residuals, weights, spec, model_type = "univariate")
}
\description{
Estimate GARCH Parameters (R Helper)
}
