% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc.R
\name{get_dcc_order}
\alias{get_dcc_order}
\title{Extract DCC order (p, q) from parameter names}
\usage{
get_dcc_order(dcc_params)
}
\arguments{
\item{dcc_params}{Named list of DCC parameters (e.g., alpha_1, alpha_2, beta_1)}
}
\value{
Named vector c(p = ..., q = ...) where q is ARCH order (number of
alpha parameters) and p is GARCH order (number of beta parameters)
}
\description{
Determines (p, q) order from parameter naming convention.
}
\keyword{internal}
