% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{hmm_bootstrap}
\alias{hmm_bootstrap}
\title{Hidden Markov Model (HMM) Bootstrap for Time Series}
\usage{
hmm_bootstrap(x, nstates = 2, num_blocks = 100, num_boots = 100)
}
\arguments{
\item{x}{Numeric vector representing the time series.}

\item{nstates}{Integer number of hidden states for the HMM.}

\item{num_blocks}{Integer number of blocks to sample for each bootstrap replicate.}

\item{num_boots}{Integer number of bootstrap replicates to generate.}
}
\value{
A list of numeric vectors, each one a bootstrap replicate.
}
\description{
Fits a Gaussian Hidden Markov Model (HMM) to a univariate time series
and generates bootstrap replicates by resampling regime-specific blocks.
}
\details{
This function:
\itemize{
\item Fits a Gaussian HMM to \code{x} using \code{depmixS4::depmix()} and \code{depmixS4::fit()}.
\item Uses Viterbi decoding (\code{posterior(fit, type = "viterbi")$state})
to assign each observation to a state.
\item Samples contiguous blocks of observations belonging to each state.
}
}
\examples{
set.seed(123)
x <- arima.sim(n = 200, list(ar = 0.5))
hmm_samples <- hmm_bootstrap(x, nstates = 2, num_blocks = 10, num_boots = 5)
length(hmm_samples)     # 5 bootstrap replicates
length(hmm_samples[[1]]) # length of one bootstrap series

}
