% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{identify_problematic_states}
\alias{identify_problematic_states}
\title{Identify Problematic States}
\usage{
identify_problematic_states(diagnostics)
}
\arguments{
\item{diagnostics}{An object of class \code{ms_diagnostics}}
}
\value{
List with problem states and issue descriptions
}
\description{
Identify states with estimation problems based on diagnostic information.
}
