% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R
\name{simulate_dcc_garch_test_data}
\alias{simulate_dcc_garch_test_data}
\title{Simulate DCC-GARCH data for testing}
\usage{
simulate_dcc_garch_test_data(
  n = 200,
  k = 2,
  alpha = 0.05,
  beta = 0.9,
  seed = 123
)
}
\arguments{
\item{n}{Number of observations}

\item{k}{Number of series}

\item{alpha}{DCC alpha parameter}

\item{beta}{DCC beta parameter}

\item{seed}{Random seed}
}
\value{
List with y (data), true_params, Qbar
}
\description{
Simulate DCC-GARCH data for testing
}
