% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbs-package.R
\docType{package}
\name{tsbs-package}
\alias{tsbs-package}
\title{tsbs: Advanced Bootstrap for Time Series}
\description{
Generates bootstrap replicates of multivariate or (multiple) univariate time
series. Supports moving and stationary block, HMM, MSVAR, MS VARMA GARCH and
wild bootstrap types.
}
\section{Main Function}{

\itemize{
\item \code{\link{tsbs}}: Main user-facing function
}
}

\seealso{
Vignettes:
\itemize{
\item \code{vignette("Diagnostics", package = "tsbs")} - User guide for
MS-VARMA-GARCH model fitting diagnostic system
}
}
\author{
\strong{Maintainer}: Martin Hoshi Vognsen \email{martinvognsen@gmail.com} (\href{https://orcid.org/0000-0003-4460-7763}{ORCID})

}
\keyword{internal}
